requirements.txt
appdirs==1.4.3
apturl==0.5.2
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.2
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.10
cachetools==5.0.0
cachy==0.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
cleo==0.8.1
click==8.0.4
clikit==0.6.2
colorama==0.4.3
comm==0.1.2
command-not-found==0.3
crashtest==0.3.1
cryptography==2.8
cupshelpers==1.0
DateTime==5.0
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.0
executing==1.2.0
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
h11==0.14.0
html5lib==1.1
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.1
ipykernel==6.21.1
ipython==8.9.0
itsdangerous==2.1.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.0.3
jupyter-client==8.0.2
jupyter-core==5.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.6
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.3
nest-asyncio==1.5.6
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
outcome==1.2.0
packaging==20.9
paramiko==2.6.0
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.8.2
platformdirs==2.5.0
poetry==1.1.13
poetry-core==1.0.7
prompt-toolkit==3.0.36
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==1.7.1
pylev==1.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.0.0
reportlab==3.5.34
requests==2.28.2
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==3.3.1
selenium==4.8.0
shellingham==1.4.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
systemd-python==234
tomlkit==0.10.0
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.0.3
wsproto==1.2.0
xkit==0.0.0
zipp==1.0.0
zope.interface==5.5.2
